refactor: move proxy accessor function to extra module
