+++ lib/file.js
@@ -6,6 +6,7 @@ const { readFileSync } = require('fs')
 const { printEnum, propertyToInlineEnumName, stringifyEnumImplementation } = require('./components/enum')
 const { normalise } = require('./components/identifier')
 const { empty } = require('./components/typescript')
+const { proxyAccessFunction } = require('./components/javascript')
 const { createObjectOf } = require('./components/wrappers')
 
 const AUTO_GEN_NOTE = '// This is an automatically generated file. Please do not change its contents manually!'
@@ -415,33 +416,7 @@ class SourceFile extends File {
         ].filter(Boolean).join('\n')
     }
     #getEntityProxyFunctionExport() {
-        return `module.exports.createEntityProxy = function(namespace, name, opts = {}) {
-    const { target, customProps } = { target: {}, customProps: [], ...opts }
-    const fq = namespace !== '' ? \`\${namespace}.\${name}\` : name
-    let _entity
-    return new Proxy(target, {
-        get: function (target, prop) {
-            // we already know the name so we skip the cds.entities proxy access
-            if (prop === "name") return fq
-            // custom properties access on 'target' as well as cached _entity property access goes here
-            if (target.hasOwnProperty(prop)) return target[prop]
-            // inline enums have to be caught here for first time access, as they do not exist on the entity
-            if (customProps.length && customProps.includes(prop)) return target[prop]
-            // last but not least we pass the property access to cds.entities
-            if (cds.entities) {
-                const entity = _entity ? _entity : (_entity = cds.entities(namespace)[name]);
-                if (!entity) throw new Error(\`'\${fq}' could not be found in cds.entities('\${namespace}')\`)
-                if (entity[prop]) target[prop] = entity[prop]
-                return target[prop]
-            }
-            throw new Error(\`Property \${prop} does not exist on entity '\${fq}' or cds.entities is not yet defined\`);
-        },
-        set: function (target, prop, newVal) {
-            target[prop] = newVal
-            return true
-        }
-    })
-}`
+        return `module.exports.createEntityProxy = ${proxyAccessFunction}`
     }
     #getEntityProxyCustomProps(name) {
         const customProps = this.entityProxies[name] ?? []
